+++ uefi/src/proto/mod.rs
@@ -31,6 +31,7 @@ pub mod shell_params;
 pub mod shim;
 pub mod string;
 pub mod tcg;
+pub mod usb;
 
 mod boot_policy;